recipe_project/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
openai==0.28.0
propcache==0.2.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
tqdm==4.66.5
urllib3==2.2.3
yarl==1.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
openai==0.28.0
propcache==0.2.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
tqdm==4.66.5
urllib3==2.2.3
yarl==1.14.0
